+++ lib/bottom_action_sheet/bottom_action_sheet_button/android_bottom_action_sheet_button.dart
@@ -5,26 +5,22 @@ class AndroidBottomActionSheetButton implements AdaptiveBottomActionSheetButton 
   @override
   Widget build(
     BuildContext context, {
-    required String text,
+    required Widget label,
     void Function(BuildContext context)? onPressed,
     required bool isDefaultAction,
     required bool isDestructiveAction,
     required bool isCancelAction,
   }) {
-    return TextButton(
-      style: TextButton.styleFrom(
-        alignment: Alignment.centerLeft,
-        textStyle: TextStyle(fontSize: 18, fontWeight: isCancelAction ? FontWeight.w600 : FontWeight.w500),
-      ),
-      onPressed: () {
-        if (onPressed != null) {
-          onPressed.call(context);
-        } else {
-          Navigator.of(context).pop();
-        }
-      },
-      child: Text(
-        text,
+    return Material(
+      child: ListTile(
+        title: label,
+        onTap: () {
+          if (onPressed != null) {
+            onPressed.call(context);
+          } else {
+            Navigator.of(context).pop();
+          }
+        },
       ),
     );
   }